Refactor TaskListPanel to be concrete and remove subclasses
